Using filter value getter
